helm mode: Make Helm chart registry configurable
Add --registry flag in helm mode install / upgrade commands to override
the default https://helm.cilium.io registry.

I intentionally avoided adding this flag to the classic mode since it's
being deprecated in the next minor release (v0.15). The behavior of the
classic mode is unchanged in this pull request.

Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>